+++ packages/vscode-extension/src/commands.ts
@@ -1,18 +1,14 @@
-import { satisfies } from 'semver';
 import vscode, { ShellExecution, Task, TaskScope } from 'vscode';
 import { GraphVisualizerView } from './graphVisualizerView';
-import { findMoonBin, getMoonVersion } from './moon';
+import type { Workspace } from './workspace';
 
 export async function checkProject(
 	project: string,
-	workspaceRoot: string,
+	workspace: Workspace,
 	modifier?: (task: Task) => void,
 ) {
-	const args = ['check', project];
-	const version = await getMoonVersion(workspaceRoot);
-
-	if (satisfies(version, '>=0.17.0 && <0.22.0')) {
-		args.push('--report');
+	if (!workspace.root || !workspace.binPath) {
+		return;
 	}
 
 	const task = new Task(
@@ -20,9 +16,18 @@ export async function checkProject(
 		TaskScope.Workspace,
 		`moon check ${project}`,
 		'moon',
-		new ShellExecution(findMoonBin(workspaceRoot)!, args, {
-			cwd: workspaceRoot,
-		}),
+		new ShellExecution(
+			workspace.binPath,
+			[
+				'check',
+				project,
+				'--log',
+				vscode.workspace.getConfiguration('moon').get('logLevel', 'info'),
+			],
+			{
+				cwd: workspace.root,
+			},
+		),
 	);
 
 	modifier?.(task);
@@ -32,14 +37,11 @@ export async function checkProject(
 
 export async function runTask(
 	target: string,
-	workspaceRoot: string,
+	workspace: Workspace,
 	modifier?: (task: Task) => void,
 ) {
-	const args = ['run', ...target.split(' ')];
-	const version = await getMoonVersion(workspaceRoot);
-
-	if (satisfies(version, '<0.22.0')) {
-		args.push('--report');
+	if (!workspace.root || !workspace.binPath) {
+		return;
 	}
 
 	const task = new Task(
@@ -47,9 +49,18 @@ export async function runTask(
 		TaskScope.Workspace,
 		`moon run ${target}`,
 		'moon',
-		new ShellExecution(findMoonBin(workspaceRoot)!, args, {
-			cwd: workspaceRoot,
-		}),
+		new ShellExecution(
+			workspace.binPath,
+			[
+				'run',
+				...target.split(' '),
+				'--log',
+				vscode.workspace.getConfiguration('moon').get('logLevel', 'info'),
+			],
+			{
+				cwd: workspace.root,
+			},
+		),
 	);
 
 	modifier?.(task);
@@ -57,21 +68,25 @@ export async function runTask(
 	await vscode.tasks.executeTask(task);
 }
 
-export async function runTaskByInput(workspaceRoot: string) {
+export async function runTaskByInput(workspace: Workspace) {
+	if (!workspace.root || !workspace.binPath) {
+		return;
+	}
+
 	const target = await vscode.window.showInputBox({
 		prompt: 'In the format of "scope:task" or ":task".',
 		title: 'Target(s)',
 	});
 
 	if (target) {
-		await runTask(target, workspaceRoot);
+		await runTask(target, workspace);
 	}
 }
 
-export async function viewActionGraph(context: vscode.ExtensionContext, workspaceRoot: string) {
-	await new GraphVisualizerView(context, workspaceRoot, 'action-graph').renderPanel();
+export async function viewActionGraph(context: vscode.ExtensionContext, workspace: Workspace) {
+	await new GraphVisualizerView(context, workspace, 'action-graph').renderPanel();
 }
 
-export async function viewProjectGraph(context: vscode.ExtensionContext, workspaceRoot: string) {
-	await new GraphVisualizerView(context, workspaceRoot, 'project-graph').renderPanel();
+export async function viewProjectGraph(context: vscode.ExtensionContext, workspace: Workspace) {
+	await new GraphVisualizerView(context, workspace, 'project-graph').renderPanel();
 }